requirements.txt
asgiref==3.10.0
Django==5.2.8
django-parler==2.3
django-autoslug==1.9.9
pillow==12.0.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
